+++ Towermap/Core/Entities/ActorManager.cs
@@ -74,43 +74,3 @@ public class ActorManager
         unusedIds.Clear();
     }
 }
-
-public class Actor 
-{
-    public string Name;
-    public Quad Texture;
-    public int Width;
-    public int Height;
-    public Vector2 Origin;
-    public ActorRender OnRender;
-    public bool ResizeableX;
-    public bool ResizeableY;
-    public Dictionary<string, object> CustomValues;
-}
-
-
-public struct ActorInfo 
-{
-    public string Name;
-    public string Texture;
-    public int Width;
-    public int Height;
-    public int OriginX;
-    public int OriginY;
-    public bool ResizeableX;
-    public bool ResizeableY;
-    public Dictionary<string, object> CustomValues;
-
-    public ActorInfo(string name, string texture, int width = 20, int height = 20, int originX = 0, int originY = 0, bool resizeableX = false, bool resizeableY = false, Dictionary<string, object> customValues = null) 
-    {
-        Name = name;
-        Texture = texture;
-        Width = width;
-        Height = height;
-        OriginX = originX;
-        OriginY = originY;
-        ResizeableX = resizeableX;
-        ResizeableY = resizeableY;
-        CustomValues = customValues;
-    }
-}
