+++ Towermap/Core/Entities/VanillaActor.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using MoonWorks.Graphics;
 using MoonWorks.Math.Float;
 using Riateu;
