+++ Towermap/Resource.cs
@@ -1,0 +1,13 @@
+using MoonWorks.Graphics;
+using MoonWorks.Graphics.Font;
+using Riateu.Graphics;
+
+namespace Towermap;
+
+public static class Resource 
+{
+    public static Texture TowerFallTexture;
+    public static TowerFallAtlas Atlas;
+    public static Quad Pixel;
+    public static Font Font;
+}
