+++ Towermap/TowerMapGame.cs
@@ -6,14 +6,6 @@ using Riateu.Graphics;
 
 namespace Towermap;
 
-public static class Resource 
-{
-    public static Texture TowerFallTexture;
-    public static TowerFallAtlas Atlas;
-    public static Quad Pixel;
-    public static Font Font;
-}
-
 public class TowermapGame : GameApp
 {
     public TowermapGame(string title, uint width, uint height, ScreenMode screenMode = ScreenMode.Windowed, bool debugMode = false) : base(title, width, height, screenMode, debugMode) {}
